requirements.txt
django
djangorestframework
django-phonenumber-field[phonenumbers]